+++ Craftsman/Builders/Tests/IntegrationTests/ProducerTestBuilder.cs
@@ -10,19 +10,20 @@
 
     public class ProducerTestBuilder
     {
-        public static void CreateTests(string testDirectory, Producer producer, string projectBaseName, IFileSystem fileSystem)
+        public static void CreateTests(string solutionDirectory, string testDirectory, string srcDirectory, Producer producer, string projectBaseName, IFileSystem fileSystem)
         {
             var classPath = ClassPathHelper.FeatureTestClassPath(testDirectory, $"{producer.ProducerName}Tests.cs", "EventHandlers", projectBaseName);
-            var fileText = WriteTestFileText(testDirectory, classPath, producer, projectBaseName);
+            var fileText = WriteTestFileText(solutionDirectory, testDirectory, srcDirectory, classPath, producer, projectBaseName);
             Utilities.CreateFile(classPath, fileText, fileSystem);
         }
 
-        private static string WriteTestFileText(string solutionDirectory, ClassPath classPath, Producer producer, string projectBaseName)
+        private static string WriteTestFileText(string solutionDirectory, string testDirectory, string srcDirectory, ClassPath classPath, Producer producer, string projectBaseName)
         {
             var testFixtureName = Utilities.GetIntegrationTestFixtureName();
-            var testUtilClassPath = ClassPathHelper.IntegrationTestUtilitiesClassPath(solutionDirectory, projectBaseName, "");
-            var producerClassPath = ClassPathHelper.ProducerFeaturesClassPath(solutionDirectory, "", producer.DomainDirectory, projectBaseName);
+            var testUtilClassPath = ClassPathHelper.IntegrationTestUtilitiesClassPath(testDirectory, projectBaseName, "");
+            var producerClassPath = ClassPathHelper.ProducerFeaturesClassPath(srcDirectory, "", producer.DomainDirectory, projectBaseName);
             
+            var messagesClassPath = ClassPathHelper.MessagesClassPath(solutionDirectory, "");
             return @$"namespace {classPath.ClassNamespace};
 
 using FluentAssertions;
@@ -30,7 +31,7 @@ using NUnit.Framework;
 using System.Threading.Tasks;
 using MassTransit;
 using MassTransit.Testing;
-using Messages;
+using {messagesClassPath.ClassNamespace};
 using Microsoft.Extensions.DependencyInjection;
 using Moq;
 using {producerClassPath.ClassNamespace};