+++ CHANGELOG.md
@@ -11,6 +11,46 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+* A `Dockerfile` and `.dockerignore` will be added to each bounded context automatically
+
+* A `docker-compose.yaml` will be added to your solution root by default for local development
+  * Just run `docker-compose up --build` in your project root
+
+  * Then set an env and apply migrations. For a postgres example:
+    * env
+
+        *Powershell*
+    
+      ```powershell
+      $Env:ASPNETCORE_ENVIRONMENT = "anything"
+      ```
+    
+        *Bash*
+    
+      ```bash
+      export ASPNETCORE_ENVIRONMENT=anything
+      ```
+    
+    * `dotnet ef database update --connection "Host=localhost;Port=3125;Database=dev_recipemanagement;Username=postgres;Password=postgres"`
+    
+  * Default settings can be overriden using a `DockerConfig` object on an `ApiTemplate`
+
+  * `SA` will always be default user for sqlserver so it can work properly
+
+  * If no ports are given for api or db, they'll be auto assigned a free port on your machine
+
+
+### Updated
+
+- Initial commit will use system git user and email as author.
+  - **DOCS:** Can be toggled off to use a generic `Craftsman` author if desired using a `UseSystemGitUser` boolean on your Domain Template
+- `Id` on `BaseEntity` is sortable and filterable by default
+
+
+## [0.13.0] - 01/27/2022
+
+### Added
+
 * Entities will now have private setters, a static `Create` method and an `Update` method to promote a DDD workflow
   * This includes base entity and the dbcontext setters for the auditable fields
 * A huge permissions overhaul
@@ -20,9 +60,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   * Role-permission mappings outside of this need tobe done manually 
   * You will be responsible for managing permissions outside of super admin
   * Policies are no longer part of the craftsman api, but a single policy will be added with each api boundary to swagger to dictate access to that boundary. It has a default of a snake case of `ProjectName`, but can be overridden
-    * **Call out in docs**. Also docs note that scope added in boundary for access should match the scope on your auth server, 
     * If using the built in auth server, this will be added for you. if not, make sure it mataches an appropriate scope in your auth server for this api boundary
-  * **Features now have a prop for `IsProtected` that, if true, will add an authorization attribute to your endpoint with the `PolicyName` and add tests that check for access**
+  * Features now have a prop for `IsProtected` that, if true, will add an authorization attribute to your endpoint with the `PolicyName` and add tests that check for access
   * Integration tests for `UserPolicyHandler`
   * Unit tests for `RolePermission`
   * `Policies` prop removed from Feature
@@ -30,6 +69,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   * Functional tests auth helper method now sets `role` instead of `scope`
   * Alice is a `SuperAdmin` and bob is a `User`
 * Added a `register:producer` command with CLI prompt
+* Added `UseSoftDelete` property to the `ApiTemplate` which is set to true. When adding an entity after the fact, Craftsman will automatically detect whether or not your project is using soft deletion by checking base entity for the appropriate property.
+* Added a `SharedKernel` project at the root to capture DTOs, exceptions, and roles (if using auth)
+* Added new `Complex` example for `new:example`
 
 
 ### Updated
@@ -67,13 +109,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
   * Producer test generated
 
-  * **TODO** docs on what a producer and consumer test look like
-
 * Minor update to naming of producer in the bus example
 
 * Default exchange type now `Fanout` for producer and consumer
 
 * Added optional (but recommended) `DomainDirectory` prop to producers and consumers. This will move them from the `EventHandlers` directory and keep them directly with features for better colocation.
+
+* Updated the 1:1 relationships to use proper scaffolding.
+
+* Updated the FK example to show proper 1:1 relationship
+
+* Entities with a Guid prop will no longer have a default of `Guid.NewGuid()`
+
+* Updated default library from NewtonSoft.Json to System.text.Json (https://github.com/pdevito3/craftsman/issues/52)
+
+* Took audit fields off of DTO
+
+* Bumped LibGit2Sharp to preview for M1 compatibility
+
 ### Fixed
 
 * Batch endpoint route updated to `batch` along with a functional testing route
@@ -82,13 +135,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Can enter null for `add:feature` batch list options where needed
 * Minor formatting fix for indentation in producers and consumers
 * Removed extra exception using from  patch integration test
-
-## [0.12.4] - 01/09/2022
-
-### Update
-
-* Updated the 1:1 relationships to use proper scaffolding.
-* Updated the FK example to show proper 1:1 relationship
+* Fixed docker image for integration tests running on macOS + M1 chip (https://github.com/pdevito3/craftsman/issues/53)
 
 ## [0.12.3] - 12/20/2021